Remove most uses of bigInt(it is super slow).

Calling NextUInt32 100,000 times from 8954.462ms to 19.119 ms(99.79% improved).

Calling NextUInt64 100,000 times from 20148.601ms to 1963.804 ms(90.25% improved).